# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "hanjan API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/beans": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Beans",
        "operationId": "list_beans_api_beans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BeanOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/beans/{bean_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Bean",
        "operationId": "get_bean_api_beans__bean_id__get",
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BeanStatsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/beans/{bean_id}/similar": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Similar Beans",
        "operationId": "get_similar_beans_api_beans__bean_id__similar_get",
        "parameters": [
          {
            "name": "bean_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Bean Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 20,
              "minimum": 1,
              "default": 5,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SimilarBeanOut"
                  },
                  "title": "Response Get Similar Beans Api Beans  Bean Id  Similar Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/brews": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Brews",
        "operationId": "list_brews_api_brews_get",
        "parameters": [
          {
            "name": "bean_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Bean Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_BrewOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/brews/{brew_id}": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Brew",
        "operationId": "get_brew_api_brews__brew_id__get",
        "parameters": [
          {
            "name": "brew_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Brew Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BrewOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/grinders/{grinder_id}/calibration": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Get Calibration",
        "operationId": "get_calibration_api_grinders__grinder_id__calibration_get",
        "parameters": [
          {
            "name": "grinder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grinder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CalibrationOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/grinders/{grinder_id}/suggest": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "Suggest Grinder Clicks",
        "operationId": "suggest_grinder_clicks_api_grinders__grinder_id__suggest_get",
        "parameters": [
          {
            "name": "grinder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Grinder Id"
            }
          },
          {
            "name": "target_d50_mm",
            "in": "query",
            "required": true,
            "schema": {
              "type": "number",
              "maximum": 3,
              "exclusiveMinimum": 0,
              "title": "Target D50 Mm"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClickSuggestionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/catalog": {
      "get": {
        "tags": [
          "public"
        ],
        "summary": "List Catalog",
        "operationId": "list_catalog_api_catalog_get",
        "parameters": [
          {
            "name": "include_unavailable",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Unavailable"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_CatalogBeanOut_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BeanOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "roaster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roaster"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "region": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "variety": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Variety"
          },
          "processing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing"
          },
          "roast_level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roast Level"
          },
          "bag_flavor_notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bag Flavor Notes"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "roaster",
          "country",
          "region",
          "variety",
          "processing",
          "roast_level",
          "bag_flavor_notes",
          "created_at"
        ],
        "title": "BeanOut"
      },
      "BeanStatsOut": {
        "properties": {
          "bean": {
            "$ref": "#/components/schemas/BeanOut"
          },
          "brew_count": {
            "type": "integer",
            "title": "Brew Count"
          },
          "avg_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Rating"
          }
        },
        "type": "object",
        "required": [
          "bean",
          "brew_count",
          "avg_rating"
        ],
        "title": "BeanStatsOut"
      },
      "BeanSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "roaster": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Roaster"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "roaster"
        ],
        "title": "BeanSummary"
      },
...
      "BrewOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "bean": {
            "$ref": "#/components/schemas/BeanSummary"
          },
          "grinder_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grinder Id"
          },
          "brewed_on": {
            "type": "string",
            "format": "date",
            "title": "Brewed On"
          },
          "grind_clicks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grind Clicks"
          },
          "water_temp_c": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water Temp C"
          },
          "dose_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dose G"
          },
          "water_g": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Water G"
          },
          "brew_time_s": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brew Time S"
          },
          "rating": {
            "type": "integer",
            "title": "Rating"
          },
          "flavor_tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flavor Tags"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "measurements": {
            "items": {
              "$ref": "#/components/schemas/GrindMeasurementOut"
            },
            "type": "array",
            "title": "Measurements"
          },
          "ratio": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ratio",
            "description": "원두 1g당 물 g.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "bean",
          "grinder_id",
          "brewed_on",
          "grind_clicks",
          "water_temp_c",
          "dose_g",
          "water_g",
          "brew_time_s",
          "rating",
          "flavor_tags",
          "note",
          "created_at",
          "measurements",
          "ratio"
        ],
        "title": "BrewOut"
      },
      "CalibrationOut": {
        "properties": {
          "grinder_id": {
            "type": "integer",
            "title": "Grinder Id"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CalibrationRowOut"
            },
            "type": "array",
            "title": "Rows"
          },
          "monotonic": {
            "type": "boolean",
            "title": "Monotonic"
          }
        },
        "type": "object",
        "required": [
          "grinder_id",
          "rows",
          "monotonic"
        ],
        "title": "CalibrationOut"
      },
      "CalibrationRowOut": {
        "properties": {
          "clicks": {
            "type": "integer",
            "title": "Clicks"
          },
          "n": {
            "type": "integer",
            "title": "N"
          },
          "d50_median_mm": {
            "type": "number",
            "title": "D50 Median Mm"
          },
          "d50_min_mm": {
            "type": "number",
            "title": "D50 Min Mm"
          },
          "d50_max_mm": {
            "type": "number",
            "title": "D50 Max Mm"
          }
        },
        "type": "object",
        "required": [
          "clicks",
          "n",
          "d50_median_mm",
          "d50_min_mm",
          "d50_max_mm"
        ],
        "title": "CalibrationRowOut"
      },
      "CatalogBeanOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "source_name": {
            "type": "string",
            "title": "Source Name"
          },
          "source_url": {
            "type": "string",
            "title": "Source Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "price": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "processing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Processing"
          },
          "normalized": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Normalized"
          },
          "first_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "First Seen At"
          },
          "last_seen_at": {
            "type": "string",
            "format": "date-time",
            "title": "Last Seen At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "source_name",
          "source_url",
          "title",
          "price",
          "available",
          "status",
          "country",
          "processing",
          "normalized",
          "first_seen_at",
          "last_seen_at"
        ],
        "title": "CatalogBeanOut"
      },
      "ClickSuggestionOut": {
        "properties": {
          "grinder_id": {
            "type": "integer",
            "title": "Grinder Id"
          },
          "target_d50_mm": {
            "type": "number",
            "title": "Target D50 Mm"
          },
          "clicks": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Clicks"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "grinder_id",
          "target_d50_mm",
          "clicks",
          "reason"
        ],
        "title": "ClickSuggestionOut"
      },
...
      "GrindMeasurementOut": {
        "properties": {
          "particle_count": {
            "type": "integer",
            "title": "Particle Count"
          },
          "mm_per_px": {
            "type": "number",
            "title": "Mm Per Px"
          },
          "d10_mm": {
            "type": "number",
            "title": "D10 Mm"
          },
          "d50_mm": {
            "type": "number",
            "title": "D50 Mm"
          },
          "d90_mm": {
            "type": "number",
            "title": "D90 Mm"
          },
          "count_d50_mm": {
            "type": "number",
            "title": "Count D50 Mm"
          },
          "min_detectable_mm": {
            "type": "number",
            "title": "Min Detectable Mm"
          },
          "clump_suspects": {
            "type": "integer",
            "title": "Clump Suspects"
          },
          "histogram": {
            "items": {
              "$ref": "#/components/schemas/HistogramBinOut"
            },
            "type": "array",
            "title": "Histogram"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Warnings"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "brew_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brew Id"
          },
          "grinder_id": {
            "type": "integer",
            "title": "Grinder Id"
          },
          "clicks": {
            "type": "integer",
            "title": "Clicks"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "particle_count",
          "mm_per_px",
          "d10_mm",
          "d50_mm",
          "d90_mm",
          "count_d50_mm",
          "min_detectable_mm",
          "clump_suspects",
          "histogram",
          "warnings",
          "id",
          "brew_id",
          "grinder_id",
          "clicks",
          "created_at"
        ],
        "title": "GrindMeasurementOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_BeanOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BeanOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[BeanOut]"
      },
      "Page_BrewOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BrewOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[BrewOut]"
      },
      "Page_CatalogBeanOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CatalogBeanOut"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[CatalogBeanOut]"
      },
...
      "SimilarBeanOut": {
        "properties": {
          "bean": {
            "$ref": "#/components/schemas/BeanOut"
          },
          "similarity": {
            "type": "number",
            "title": "Similarity"
          },
          "brew_count": {
            "type": "integer",
            "title": "Brew Count"
          },
          "avg_rating": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Rating"
          }
        },
        "type": "object",
        "required": [
          "bean",
          "similarity",
          "brew_count",
          "avg_rating"
        ],
        "title": "SimilarBeanOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}